pyproject.toml
[tool.poetry]
name = "Notti bot"
version = "0.1.0"
description = ""
authors = ["Dmytro Filin <dmitriy.fillin@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytelegrambotapi = "^4.14.1"
selenium = "^4.16.0"
webdriver-manager = "^4.0.1"
python-dotenv = "^1.0.0"
python-telegram-bot = "^20.7"
schedule = "^1.2.1"
flask = "^3.0.0"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.2"
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
